move parser and saver into folder 'lib'
